feat:deploy

+++ routes/courses.js
@@ -2,15 +2,28 @@ const { Router } = require("express");
 const Course = require("../models/course");
 const router = Router();
 const auth = require("../middleware/auth");
+const { courseValidators } = require("../utils/validators");
+const { validationResult } = require("express-validator");
+
+function isOwner(course, req) {
+  return course.userId.toString() === req.user._id.toString();
+}
 
 router.get("/", async (req, res) => {
-  const courses = await Course.find().populate("userId");
+  try {
+    const courses = await Course.find()
+      .populate("userId", "email name")
+      .select("price title img");
 
-  res.render("courses", {
-    title: "courses",
-    isCourses: true,
-    courses,
-  });
+    res.render("courses", {
+      title: "courses",
+      isCourses: true,
+      userId: req.user ? req.user._id.toString() : null,
+      courses,
+    });
+  } catch (e) {
+    console.log(e);
+  }
 });
 
 router.get("/:id/edit", auth, async (req, res) => {
@@ -18,35 +31,65 @@ router.get("/:id/edit", auth, async (req, res) => {
     return res.redirect("/");
   }
 
-  const course = await Course.findById(req.params.id);
+  try {
+    const course = await Course.findById(req.params.id);
 
-  res.render("courseEdit", {
-    title: `Edit ${course.title}`,
-    course,
-  });
+    if (!isOwner(course, req)) {
+      return res.redirect("/courses");
+    }
+
+    res.render("courseEdit", {
+      title: `Edit ${course.title}`,
+      course,
+    });
+  } catch (e) {
+    console.log(e);
+  }
 });
 
 router.get("/:id", async (req, res) => {
-  const course = await Course.findById(req.params.id);
-  res.render("course", {
-    layout: "empty",
-    title: `Course ${course.title}`,
-    course,
-  });
+  try {
+    const course = await Course.findById(req.params.id);
+    res.render("course", {
+      layout: "empty",
+      title: `Course ${course.title}`,
+      course,
+    });
+  } catch (e) {
+    console.log(e);
+  }
 });
 
 router.post("/remove", auth, async (req, res) => {
   try {
-    await Course.deleteOne({ _id: req.body.id });
+    await Course.deleteOne({ _id: req.body.id, userId: req.user._id });
     res.redirect("/courses");
   } catch (e) {
     console.log(e);
   }
 });
 
-router.post("/edit", auth, async (req, res) => {
-  await Course.findByIdAndUpdate(req.body.id, req.body);
-  res.redirect("/courses");
+router.post("/edit", auth, courseValidators, async (req, res) => {
+  const errors = validationResult(req);
+  const { id } = req.body;
+
+  if (!errors.isEmpty()) {
+    return res.status(422).redirect(`/courses/${id}/edit?allow=true`);
+  }
+
+  try {
+    delete req.body.id;
+    const course = await Course.findById(id);
+    if (!isOwner(course, req)) {
+      return res.redirect("/courses");
+    }
+    Object.assign(course, req.body);
+    await course.save();
+
+    res.redirect("/courses");
+  } catch (e) {
+    console.log(e);
+  }
 });
 
 module.exports = router;
